busmanagement/tourweb/requirements.txt
Django==4.1.7
django_rest_swagger==2.2.0
djangorestframework==3.14.0
drf_yasg==1.21.5
